chore: extracted promiseTimeout to utils file
