+++ src/modules/setup.ts
@@ -2,6 +2,7 @@ import Multiaddr from 'multiaddr'
 import mafmt from 'mafmt'
 
 import logger from '../logger'
+import { promiseTimeout } from './utils'
 
 const multiaddrformats = [
   mafmt.P2P,
@@ -19,28 +20,6 @@ const multiaddrformats = [
  */
 export function multiaddrValidator (multiaddr: string) {
   return multiaddrformats.some((validator) => validator.matches(multiaddr))
-}
-
-/**
- * Apply a timeout to a given promise.
- *
- * @param ms - timeout in milliseconds
- * @param promise - Promise to fulfill or reject
- *
- * @returns Promise<any>
- */
-function promiseTimeout (ms: number, promise: Promise<any>) {
-  const timeout = new Promise((resolve, reject) => {
-    const id = setTimeout(() => {
-      clearTimeout(id)
-      reject(new Error(`Timed out in ${ms}ms`))
-    }, ms)
-  })
-
-  return Promise.race([
-    promise,
-    timeout
-  ])
 }
 
 /**
